+++ lib/cupertino_date_selector.dart
@@ -1,2 +1,5 @@
 export 'src/cupertino_date_selector_widget.dart';
 export 'src/cupertino_date_selector_theme.dart';
+export 'src/cupertino_date_selector_selection_mode.dart';
+export 'src/cupertino_date_selector_modifiers.dart';
+export 'src/cupertino_date_selector_controller.dart';